backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
scikit-learn==1.5.1
nltk==3.9.1
pandas==2.2.2
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.2
joblib==1.4.2
